feat: 심화과제 구현
