+++ src/context/NotificationContext/NotificationProvider.tsx
@@ -1,0 +1,43 @@
+import { useState } from "react";
+import { Notification } from "../../type";
+import { useCallback, useMemo } from "../../@lib";
+import { NotificationContext } from "./NotificationContext";
+
+const NotificationProvider = ({ children }: { children: React.ReactNode }) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback(
+    (message: string, type: Notification["type"]) => {
+      const newNotification: Notification = {
+        id: Date.now(),
+        message,
+        type,
+      };
+      setNotifications((prev) => [...prev, newNotification]);
+    },
+    [],
+  );
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) =>
+      prev.filter((notification) => notification.id !== id),
+    );
+  }, []);
+
+  const notificationValue = useMemo(
+    () => ({
+      notifications,
+      addNotification,
+      removeNotification,
+    }),
+    [notifications, addNotification, removeNotification],
+  );
+
+  return (
+    <NotificationContext.Provider value={notificationValue}>
+      {children}
+    </NotificationContext.Provider>
+  );
+};
+
+export default NotificationProvider;
