+++ src/context/NotificationContext/type.ts
@@ -1,0 +1,7 @@
+import { Notification } from "../../type";
+
+export interface NotificationContextType {
+  notifications: Notification[];
+  addNotification: (message: string, type: Notification["type"]) => void;
+  removeNotification: (id: number) => void;
+}
