+++ src/context/ThemeContext/type.ts
@@ -1,0 +1,4 @@
+export interface ThemeContextType {
+  theme: string;
+  toggleTheme: () => void;
+}
