+++ src/context/UserContext/type.ts
@@ -1,0 +1,7 @@
+import { User } from "../../type";
+
+export interface UserContextType {
+  user: User | null;
+  login: (email: string, password: string) => void;
+  logout: () => void;
+}
