+++ src/type/index.ts
@@ -1,0 +1,3 @@
+export type { Notification } from "./notification";
+export type { User } from "./user";
+export type { Item } from "./item";
