+++ src/type/item.ts
@@ -1,0 +1,6 @@
+export interface Item {
+  id: number;
+  name: string;
+  category: string;
+  price: number;
+}
